# workers/collaborative-state/docs/openapi.yaml
openapi: 3.0.3
info:
  title: P1 API
  description: |
    REST API for P1 — a JSON state versioning and real-time collaboration platform.

    ## Overview

    P1 provides:
    - **Version Control**: Git-like branching and merging for JSON documents
    - **Real-time Collaboration**: WebSocket-based collaborative editing with presence awareness
    - **Agent Coordination**: AI agent support with human-first editing priority
    - **Presence Awareness**: Real-time visibility of who is editing what

    ## Authentication

    All authenticated endpoints require one of:
    - **Bearer Token**: `Authorization: Bearer <jwt-token>`
    - **API Key**: `X-API-Key: <api-key>` header or `?apiKey=<key>` query parameter

    ## Rate Limits

    - Standard endpoints: 1000 requests/minute per principal
    - Real-time connections: 100 concurrent WebSocket connections per document

    ## Error Responses

    All errors follow a consistent format:
    ```json
    {
      "error": "Error message",
      "details": {
        "reason": "Specific reason",
        "suggestions": ["Possible fix"]
      }
    }
    ```

  version: 1.0.0
  contact:
    name: Pantheon Engineering
    url: https://pantheon.io
  license:
    name: Proprietary
    url: https://pantheon.io/terms
servers:
  - url: https://ccr.p1.pantheon.io
    description: Production
  - url: http://localhost:8787
    description: Local Development
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
paths:
...
  /api/sites/{siteId}/branches/{branchId}/datasources:
    parameters:
      - $ref: '#/components/parameters/siteIdParam'
      - $ref: '#/components/parameters/branchIdParam'
    get:
      summary: List datasources
      description: |
        Returns all datasources registered on a branch. Datasources are
        auto-generated when a template is created (one datasource per template)
        and can also be deleted manually.
      operationId: listDatasources
      tags: [Datasources]
      responses:
        '200':
          description: List of datasources
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasources:
                    type: array
                    items:
                      $ref: '#/components/schemas/Datasource'
              example:
                datasources:
                  - type: local
                    name: blog-post
                    templateName: Blog Post
                    templateId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    autoGenerated: true
                    version: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/sites/{siteId}/branches/{branchId}/datasources/{datasourceName}:
    parameters:
      - $ref: '#/components/parameters/siteIdParam'
      - $ref: '#/components/parameters/branchIdParam'
      - name: datasourceName
        in: path
        required: true
        description: Datasource name (kebab-case slug matching the template name)
        schema:
          type: string
          example: blog-post
    get:
      summary: Get datasource by name
      description: Returns a single datasource definition.
      operationId: getDatasource
      tags: [Datasources]
      responses:
        '200':
          description: Datasource details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datasource'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a datasource
      description: |
        Permanently removes a datasource. Requires `canEdit` permission on the branch.
        Any queries referencing this datasource will fail until re-pointed.
      operationId: deleteDatasource
      tags: [Datasources]
      responses:
        '200':
          description: Datasource deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # Queries
  # ===========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT/OAuth token from the CSS OAuth 2.0 flow (/auth/authorize)

    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for agent authentication

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    siteIdParam:
      name: siteId
      in: path
      required: true
      description: Site UUID
      schema:
        type: string
        format: uuid
    branchIdParam:
      name: branchId
      in: path
      required: true
      description: Branch UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: object
          properties:
            reason:
              type: string
            suggestions:
              type: array
              items:
                type: string
...
    Datasource:
      type: object
      properties:
        type:
          type: string
          enum: [local]
          description: Datasource type. Currently only `local` (template-backed) is supported.
        name:
          type: string
          description: Unique name (kebab-case slug derived from the template name).
          example: blog-post
        description:
          type: string
          nullable: true
        templateName:
          type: string
          description: Human-readable name of the linked template.
          example: Blog Post
        templateId:
          type: string
          format: uuid
          description: UUID of the template this datasource references.
        autoGenerated:
          type: boolean
          description: Whether this datasource was auto-created by a template hook.
        version:
          type: integer
          description: Internal version counter for the datasource document.

    # -------------------------------------------------------------------------
    # Queries
    # -------------------------------------------------------------------------
...